mcp_servers/azure_function/requirements.txt
azure-functions==1.21.3
mcp>=0.1.0
httpx>=0.27.0
PyJWT>=2.9.0
jwcrypto>=1.5.6
python-dotenv>=1.0.1
pydantic>=2.6.4
structlog>=24.1.0
tenacity>=8.3.0
